finished adding user authentication, JWT functionality, and testing

+++ ex4/BlogList/controllers/blogs.js
@@ -1,7 +1,7 @@
+const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
-const { checkBlogAttributes } = require('../utils/middleware');
+const { setDefaultBlogFields, userExtractor } = require('../utils/middleware');
 
 blogsRouter.get('/', async (req, res) => {
 	const blogsList = await Blog.find({}).populate('user', {
@@ -11,8 +11,16 @@ blogsRouter.get('/', async (req, res) => {
 	res.json(blogsList);
 });
 
-blogsRouter.delete('/:id', async (req, res, next) => {
+blogsRouter.delete('/:id', userExtractor, async (req, res, next) => {
 	try {
+		const blog = await Blog.findById(req.params.id);
+
+		if (blog.user.toString() !== req.user._id.toString()) {
+			return res
+				.status(401)
+				.send({ error: 'user does not have access to this resource' });
+		}
+
 		await Blog.findByIdAndDelete(req.params.id);
 		res.status(204).end();
 	} catch (err) {
@@ -20,53 +28,62 @@ blogsRouter.delete('/:id', async (req, res, next) => {
 	}
 });
 
-blogsRouter.post('/', checkBlogAttributes, async (req, res) => {
-	const body = req.body;
-	const user = await User.findById(req.body.userId);
-	if (!user) {
-		return res.status(404).json({ error: 'userId missing or not valid' });
+blogsRouter.post(
+	'/',
+	setDefaultBlogFields,
+	userExtractor,
+	async (req, res, next) => {
+		try {
+			const body = req.body;
+			const blog = new Blog({
+				title: body.title,
+				author: body.author,
+				user: req.user._id,
+				url: body.url,
+				likes: body.likes,
+			});
+
+			const savedBlog = await blog.save();
+
+			req.user.blogs = req.user.blogs.concat(savedBlog._id);
+			await req.user.save();
+
+			res.status(201).json(savedBlog);
+		} catch (err) {
+			next(err);
+		}
 	}
+);
 
-	if (body.likes === undefined) {
-		body.likes = 0;
+blogsRouter.put(
+	'/:id',
+	setDefaultBlogFields,
+	userExtractor,
+	async (req, res, next) => {
+		try {
+			const { author, title, url, likes } = req.body;
+			const originalBlog = await Blog.findById(req.params.id);
+			if (!originalBlog) {
+				return res.status(404).json({ error: 'blog not found' });
+			}
+
+			if (originalBlog.user.toString() !== req.user._id.toString()) {
+				return res
+					.status(401)
+					.json({ error: 'user does not have access to this resource' });
+			}
+			originalBlog.author = author;
+			originalBlog.title = title;
+			originalBlog.url = url;
+			originalBlog.likes = likes;
+			originalBlog.user = req.user._id;
+
+			const updatedBlog = await originalBlog.save();
+			res.status(200).json(updatedBlog);
+		} catch (err) {
+			next(err);
+		}
 	}
-
-	const blog = new Blog({
-		title: body.title,
-		author: body.author,
-		user: user._id,
-		url: body.url,
-		likes: body.likes,
-	});
-
-	const savedBlog = await blog.save();
-
-	user.blogs = user.blogs.concat(savedBlog._id);
-	await user.save();
-
-	res.status(201).json(savedBlog);
-});
-
-blogsRouter.put('/:id', checkBlogAttributes, async (req, res) => {
-	const { author, title, url, likes, userId } = req.body;
-	const originalBlog = await Blog.findById(req.params.id);
-	if (!originalBlog) {
-		return res.status(404).json({ error: 'blog not found' });
-	}
-
-	const user = await User.findById(userId);
-	if (!user) {
-		return res.status(404).json({ error: 'userId missing or not valid' });
-	}
-
-	originalBlog.author = author;
-	originalBlog.title = title;
-	originalBlog.url = url;
-	originalBlog.likes = likes;
-	originalBlog.user = user._id;
-
-	const updatedBlog = await originalBlog.save();
-	res.status(200).json(updatedBlog);
-});
+);
 
 module.exports = blogsRouter;
